# swagger-openapi/product-openapi.yaml
openapi: 3.0.3
components:
  schemas:
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    ProductDto:
      type: object
      required:
      - name
      - price
      properties:
        price:
          type: number
        name:
          type: string
          minLength: 1
        description:
          type: string
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
        messages:
          type: string
        details:
          type: array
          items: {
            type: string
          }
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        createdAt:
          type: string          
paths:
  /api/products:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json: {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
      summary: Get Products
      tags:
      - Product Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDto"
        required: true
      responses:
        "201":
          description: OK
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            } 
        "400":
          description: Product already exists with name
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/ErrorResponse"
              }
            }            
      summary: Add Product
      tags:
      - Product Resource
  /api/products/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: description
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
        "400":
          description: Product not found with id
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/ErrorResponse"
              }
            }             
      summary: Update Product
      tags:
      - Product Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
        "400":
          description: Product not found with id
          content:
            application/json: {
              "schema": {
                "$ref": "#/components/schemas/ErrorResponse"
              }
            } 
      summary: Get Product By Id
      tags:
      - Product Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: No Content
          content:
            application/json: {}
        "404":
          description: Not Found
          content:
            application/json: {}            
      summary: Delete Product
      tags:
      - Product Resource
info:
  title: product-system API
  version: 1.0.0-SNAPSHOT